curriculum config refactoring
The separated CurriculumConfig allows the CMS to use that config
independently from the AppConfigModel. The AppConfigModel requires
several other files.
