+++ cms/src/preview-link-control.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import { CmsWidgetControlProps } from "netlify-cms-core";
 
-import { getGuideJson, getUnitJson } from "../../src/models/curriculum/unit";
+import { getGuideJson, getUnitJson } from "../../src/models/curriculum/unit-utils";
 import { DocumentModelType } from "../../src/models/document/document";
 import { stripPTNumberFromBranch } from "../../src/utilities/branch-utils";
 import { urlParams } from "../../src/utilities/url-params";
-import { AppConfigModel } from "../../src/models/stores/app-config-model";
-import { appConfigSnapshot } from "../../src/app-config";
 import { defaultCurriculumBranch } from "./cms-constants";
+import { CurriculumConfig } from "../../src/models/stores/curriculum-config";
+
+import curriculumConfigJson from "../../src/clue/curriculum-config.json";
 
 import "./custom-control.scss";
 import "./preview-link-control.scss";
@@ -49,14 +50,14 @@ export class PreviewLinkControl extends React.Component<CmsWidgetControlProps, I
     }
     this.unit = urlParams.unit ?? this.pathParts?.[1] ?? defaultUnit;
 
-    const appConfig = AppConfigModel.create(appConfigSnapshot);
+    const curriculumConfig = CurriculumConfig.create(curriculumConfigJson, urlParams);
 
     // Finish setting up the preview link after reading the unit json
     this.isTeacherGuide = this.pathParts?.[2] === "teacher-guide";
     if (this.isTeacherGuide) {
-      getGuideJson(this.unit, appConfig).then((unitJson: DocumentModelType) => this.setPreviewLink(unitJson));
+      getGuideJson(this.unit, curriculumConfig).then((unitJson: DocumentModelType) => this.setPreviewLink(unitJson));
     } else {
-      getUnitJson(this.unit, appConfig).then((unitJson: DocumentModelType) => this.setPreviewLink(unitJson));
+      getUnitJson(this.unit, curriculumConfig).then((unitJson: DocumentModelType) => this.setPreviewLink(unitJson));
     }
 
     this.state = {
