+++ src/clue/curriculum-config.json
@@ -1,0 +1,40 @@
+{
+  "curriculumBaseUrl": "https://models-resources.concord.org/clue-curriculum",
+  "unitCodeMap": {
+    "bio4community": "bio",
+    "completely-rational": "cr",
+    "comparing-and-scaling": "cas",
+    "dataflow-example": "dfe",
+    "example-curriculum": "example",
+    "example-curriculum/example-no-group-share": "example-no-group-share",
+    "example-curriculum/example-no-my-work": "example-no-my-work",
+    "example-curriculum/example-show-nav-panel": "example-show-nav-panel",
+    "example-curriculum/example-config-subtabs": "example-config-subtabs",
+    "example-curriculum/example-no-section-problem-tab": "example-no-section-problem-tab",
+    "example-curriculum/example-template": "example-template",
+    "example-curriculum/example-variables": "example-variables",
+    "example-curriculum/example-shared-data": "example-shared-data",
+    "filling-and-wrapping": "faw",
+    "frogs-fleas-and-painted-cubes": "ffpc",
+    "flip-spin-slide-stretch": "fsss",
+    "growing-growing-growing": "ggg",
+    "how-likely-is-it": "hlit",
+    "help": "help",
+    "looking-for-pythagoras": "lfp",
+    "m2studio": "m2s",
+    "m2s/m2studio-prototype": "m2s-proto",
+    "mothed": "moth",
+    "mods": "mods",
+    "moving-straight-ahead": "msa",
+    "neural-engineering": "brain",
+    "pc2": "pc",
+    "s s": "sas",
+    "s+s": "sas",
+    "shapes-and-designs": "sad",
+    "samples-and-populations": "sap",
+    "stretching-and-shrinking": "sas",
+    "say-it-with-symbols": "sws",
+    "thinking-with-math-models": "twmm"
+  },
+  "defaultUnit": "sas"
+}
