+++ src/components/tiles/tile-component.tsx
@@ -324,8 +324,8 @@ export class TileComponent extends BaseComponent<IProps, IState> {
   };
 
   private handleCopyImportJsonToClipboard = () => {
-    const { appConfig, unit } = this.stores;
-    const unitBasePath = appConfig.getUnitBasePath(unit.code);
+    const { curriculumConfig, unit } = this.stores;
+    const unitBasePath = curriculumConfig.getUnitBasePath(unit.code);
     const transformImageUrl = (url?: string, filename?: string) => {
       return transformCurriculumImageUrl(url, unitBasePath, filename);
     };
