+++ src/hooks/use-stores.test.ts
@@ -36,7 +36,7 @@ describe("useStores", () => {
   describe("simple store hooks", () => {
     beforeEach(() => resetMocks());
     it("should return the requested store", () => {
-      const appConfig = AppConfigModel.create({ curriculumBaseUrl: "https://curriculum.example.com", config: unitConfigDefaults });
+      const appConfig = AppConfigModel.create({ config: unitConfigDefaults });
       const _class = ClassModel.create({ name: "Class 1", classHash: "hash-1" });
       const demo = DemoModel.create({ class: DemoClassModel.create({ id: "class-1", name: "Class 1" }) });
       const groups = GroupsModel.create();
