+++ src/lib/portal-api.test.ts
@@ -1,9 +1,9 @@
 import nock from "nock";
-import { getPortalClassOfferings, getPortalOfferings, PortalOfferingParser } from "./portal-api";
+import { getPortalClassOfferings, getPortalOfferings } from "./portal-api";
 import { IPortalOffering } from "./portal-types";
 import { TeacherMineClasses, TeacherOfferings } from "../test-fixtures/sample-portal-offerings";
 import { AppConfigModelType } from "../models/stores/app-config-model";
-import { specAppConfig } from "../models/stores/spec-app-config";
+import { CurriculumConfig } from "../models/stores/curriculum-config";
 
 const userType = "teacher";
 const userID = 22;
@@ -38,70 +38,13 @@ describe("Portal Offerings", () => {
     });
   });
 
-  describe("PortalOfferingParser", () => {
-    const { getProblemOrdinal, getUnitCode } = PortalOfferingParser;
-
-    const appConfig = specAppConfig();
-    const samplePortalOffering = {
-      id: 1190,
-      teacher: "Dave Love",
-      clazz: "ClueClass1",
-      clazz_id: 242,
-      activity: "CLUE 1.2: Stretching a Figure - Comparing Similar Figures",
-      activity_url: "https://collaborative-learning.concord.org/branch/master/?problem=1.2&unit=foo"
-    };
-
-    describe("getProblemOrdinal", () => {
-      it("should return a problemOrdinal", () => {
-        const ordinal = getProblemOrdinal(samplePortalOffering.activity_url);
-        expect(ordinal).toEqual("1.2");
-      });
-    });
-
-    describe("getUnitCode", () => {
-      it("should return a unit code for problem", () => {
-        const unitCode = getUnitCode(samplePortalOffering.activity_url, appConfig);
-        expect(unitCode).toEqual("foo");
-      });
-
-      it("should return a mapped unit code for legacy units", () => {
-        const barAppConfig = specAppConfig({ unitCodeMap: { foo: "bar" } as any });
-        const unitCode = getUnitCode(samplePortalOffering.activity_url, barAppConfig);
-        expect(unitCode).toEqual("bar");
-      });
-    });
-  });
-
   describe("PortalOfferingParserWithDefaults", () => {
-    const { getProblemOrdinal, getUnitCode } = PortalOfferingParser;
-
-    const samplePortalOffering = {
-      id: 1190,
-      teacher: "Dave Love",
-      clazz: "ClueClass1",
-      clazz_id: 242,
-      activity: "CLUE 1.2: Stretching a Figure - Comparing Similar Figures",
-      activity_url: "https://collaborative-learning.concord.org/branch/master/"
-    };
-
-    describe("getProblemOrdinal", () => {
-      it(`should default to 'undefined'`, () => {
-        const ordinal = getProblemOrdinal(samplePortalOffering.activity_url);
-        expect(ordinal).toBeUndefined();
-      });
-    });
-
-    describe("getUnitCode", () => {
-      const appConfig = specAppConfig();
-      it(`should default to 'undefined'`, () => {
-        const unitCode = getUnitCode(samplePortalOffering.activity_url, appConfig);
-        expect(unitCode).toBeUndefined();
-      });
-    });
+    const curriculumConfig = CurriculumConfig.create({curriculumBaseUrl: ""});
 
     describe("getPortalClassOfferings", () => {
+      // FIXME: I'd guess this is broken
       const mockAppConfig = {
-        defaultUnit: "sas", unitCodeMap: {}, config: { defaultProblemOrdinal: "1.1" }
+        config: { defaultProblemOrdinal: "1.1" }
       } as AppConfigModelType;
       const mockUrlParams = {
               class: "https://learn.staging.concord.org/api/v1/classes/242",
@@ -109,7 +52,7 @@ describe("Portal Offerings", () => {
               reportType: "report-type",
               token: "token"
             };
-      const offerings = getPortalClassOfferings(TeacherOfferings, mockAppConfig, mockUrlParams);
+      const offerings = getPortalClassOfferings(TeacherOfferings, mockAppConfig, curriculumConfig, mockUrlParams);
       // TeacherOfferings has one non-CLUE activity
       expect(offerings.length).toBe(TeacherOfferings.length - 1);
     });
