+++ src/models/curriculum/unit.ts
@@ -11,7 +11,6 @@ import {
 import { ProblemModelType } from "./problem";
 import { resumeSupportContentParsing, SupportModel, suspendSupportContentParsing } from "./support";
 import { StampModel } from "../../plugins/drawing/model/stamp";
-import { AppConfigModelType } from "../stores/app-config-model";
 import { NavTabsConfigModel } from "../stores/nav-tabs";
 import { SettingsMstType } from "../stores/settings";
 import { UnitConfiguration } from "../stores/unit-configuration";
@@ -141,108 +140,6 @@ export const UnitModel = types.snapshotProcessor(ModernUnitModel, {
   }
 });
 export interface UnitModelType extends Instance<typeof UnitModel> {}
-
-function getUnitSpec(unitId: string | undefined, appConfig: AppConfigModelType) {
-  const requestedUnit = unitId ? appConfig.getUnit(unitId) : undefined;
-  if (unitId && !requestedUnit) {
-    console.warn(`unitId "${unitId}" not found in appConfig.units`);
-  }
-  return requestedUnit || (appConfig.defaultUnit ? appConfig.getUnit(appConfig.defaultUnit) : undefined);
-}
-
-const getExternalProblemSectionData =
-  async (invIdx: number, probIdx: number, sectIdx: number, dataUrl: string, sectionPath?: string) => {
-    try {
-      const sectionData = await fetch(dataUrl).then(res => res.json());
-      sectionData.sectionPath = sectionPath;
-      return { invIdx, probIdx, sectIdx, sectionData };
-    } catch (error) {
-      throw new Error(`Failed to load problem-section ${dataUrl} cause:\n ${error}`);
-    }
-};
-
-const populateProblemSections = async (content: Record<string, any>, unitUrl: string) => {
-  const externalSectionsArray = [];
-  for (let invIdx = 0; invIdx < content.investigations.length; invIdx++) {
-    const investigation = content.investigations[invIdx];
-    for (let probIdx = 0; probIdx < investigation.problems.length; probIdx++) {
-      const problem = investigation.problems[probIdx];
-      for (let sectIdx = 0; sectIdx < problem.sections.length; sectIdx++) {
-        // Currently, curriculum files can either contain their problem section data inline
-        // or in external JSON files. In the latter case, the problem sections arrays will
-        // be made up of strings that are paths to the external files. We fetch the data from
-        // those files and populate the section with it. Otherwise, we leave the section as
-        // is. Eventually, all curriculum files will be updated so their problem section data
-        // is in external files.
-        const section = problem.sections[sectIdx];
-        if (typeof section === "string") {
-          const sectionDataFile = section;
-          const sectionDataUrl = new URL(sectionDataFile, unitUrl).href;
-          externalSectionsArray.push(
-            getExternalProblemSectionData(invIdx, probIdx, sectIdx, sectionDataUrl, sectionDataFile)
-          );
-        }
-      }
-    }
-  }
-  if (externalSectionsArray.length > 0) {
-    await Promise.all(externalSectionsArray).then((sections: any) => {
-      for (const section of sections) {
-        const { invIdx, probIdx, sectIdx, sectionData } = section;
-        content.investigations[invIdx].problems[probIdx].sections[sectIdx] = sectionData;
-      }
-    });
-  }
-  return content;
-};
-
-export function getUnitJson(unitId: string | undefined, appConfig: AppConfigModelType) {
-  const unitSpec = getUnitSpec(unitId, appConfig);
-  const unitUrl = unitSpec?.content;
-  return fetch(unitUrl!)
-           .then(async response => {
-             if (response.ok) {
-               const unitContent = await response.json();
-               const fullUnitContent = unitContent && populateProblemSections(unitContent, unitUrl!);
-               return fullUnitContent;
-             } else {
-               // If the unit content is not found, return the response so that the caller can
-               // handle it appropriately.
-               if (response.status === 404) {
-                 return response;
-               } else {
-                 throw Error(`Request rejected with status ${response.status}`);
-               }
-             }
-           })
-           .catch(error => {
-             throw Error(`Failed to load unit ${unitUrl} cause:\n ${error}`);
-           });
-}
-
-export function getGuideJson(unitId: string | undefined, appConfig: AppConfigModelType) {
-  const unitSpec = getUnitSpec(unitId, appConfig);
-  const guideUrl = unitSpec?.guide;
-  return fetch(guideUrl!)
-          .then(async response => {
-            if (response.ok) {
-              const guideContent = await response.json();
-              const fullGuideContent = guideContent && populateProblemSections(guideContent, guideUrl!);
-              return fullGuideContent;
-            } else {
-              // If the guide content is not found, return the response so that the caller can
-              // handle it appropriately.
-              if (response.status === 404) {
-                return response;
-              } else {
-                throw Error(`Request rejected with status ${response.status}`);
-              }
-            }
-          })
-          .catch(error => {
-            throw Error(`Request rejected with exception: ${error}`);
-          });
-}
 
 export function createUnitWithoutContent(unitJson: any) {
   // read the unit content, but don't instantiate section contents (DocumentModels)
