+++ src/models/stores/app-config-model.test.ts
@@ -9,7 +9,7 @@ describe("ConfigurationManager", () => {
   const keys = Object.keys(unitConfigDefaults).filter(prop => !excludeProps.includes(prop)) as SimpleProps[];
 
   it("can be constructed with just unitConfigDefaults and return those unitConfigDefaults", () => {
-    const appConfig = AppConfigModel.create({ curriculumBaseUrl: "https://curriculum.example.com", config: unitConfigDefaults });
+    const appConfig = AppConfigModel.create({ config: unitConfigDefaults });
     keys.forEach((prop: SimpleProps) => {
       expect(appConfig[prop]).toEqual(unitConfigDefaults[prop]);
     });
@@ -19,7 +19,7 @@ describe("ConfigurationManager", () => {
   });
 
   it("can be constructed with unitConfigDefaults and unitConfigOverrides and return the unitConfigOverrides", () => {
-    const appConfig = AppConfigModel.create({ curriculumBaseUrl: "https://curriculum.example.com", config: unitConfigDefaults });
+    const appConfig = AppConfigModel.create({ config: unitConfigDefaults });
     appConfig.setConfigs([unitConfigOverrides]);
     keys.forEach((prop: SimpleProps) => {
       if (prop === "disabledFeatures") {
@@ -39,37 +39,8 @@ describe("ConfigurationManager", () => {
     expect(appConfig.isFeatureSupported("baz")).toBe(true);
   });
 
-  describe("unit handling", () => {
-    const getAppConfig = () => AppConfigModel.create({
-      curriculumBaseUrl: "https://curriculum.example.com",
-      config: unitConfigDefaults
-    });
-
-    it("can get URLs for remote curriculum content from a unit code", () => {
-      const appConfig = getAppConfig();
-      const exampleUnitCode = "example-unit-code";
-      const exampleUnit = {
-        "content": `https://curriculum.example.com/branch/main/${exampleUnitCode}/content.json`,
-        "guide": `https://curriculum.example.com/branch/main/${exampleUnitCode}/teacher-guide/content.json`
-      };
-      expect(appConfig.getUnit(exampleUnitCode)).toStrictEqual(exampleUnit);
-      expect(appConfig.getUnitBasePath(exampleUnitCode)).toBe(exampleUnitCode);
-    });
-
-    it("can get URLs for remote curriculum content from a unit URL", () => {
-      const appConfig = getAppConfig();
-      const exampleUnitUrl = "https://concord.org/content.json";
-      const exampleUnit = {
-        "content": "https://concord.org/content.json",
-        "guide": "https://concord.org/teacher-guide/content.json"
-      };
-      expect(appConfig.getUnit(exampleUnitUrl)).toStrictEqual(exampleUnit);
-      // FIXME: This is probably a bug because this base path is used to compute the location of some images
-      // by tacking on an /images/... to this base path.
-      // With the current code it means the computed URL would be something like:
-      //   https://concord.org/content.json/images/...
-      expect(appConfig.getUnitBasePath(exampleUnitUrl)).toBe(exampleUnitUrl);
-    });
-
+  it("can be constructed with no props", () => {
+    const model = AppConfigModel.create();
+    expect(model.config).toBeUndefined();
   });
 });
